refactor(pipeline/network): drain AuthDiscovery.ts grandfather (Phase 8.5a commit 2)
Phase 8.5a / Commit 2 of the Network canonical-10 drain.

- Split AuthDiscovery.ts (~425 raw / ~260 eff LoC, 11 funcs over 10 LoC) into
  8 focused sub-modules under Mediator/Network/AuthDiscovery/:
  - Tokens.ts (67 LoC) — TOKEN_BODY_FIELDS, STORAGE_AUTH_KEYS,
    AUTH_HEADER_NAMES, AUTH_POLL_TIMEOUT, AUTH_POLL_INTERVAL constants
    + IStorageAuth shape + PrefixedAuthToken branded type +
    prefixToken / tryParseJsonToken / extractFromParsed helpers
  - HeadersTier.ts (31 LoC) — Tier 1: extractAuthHeader +
    discoverFromHeaders
  - ResponseTier.ts (60 LoC) — Tier 2: findTokenInFlat +
    searchBodyForToken + filterAuthEndpoints + discoverFromResponses
  - StorageMain.ts (45 LoC) — Tier 3a: readMainStorageRaw +
    pickTokenFromRaw + discoverFromStorage
  - StorageFrames.ts (132 LoC) — Tier 3b: tokenFromJsonValue +
    tokenFromLongRaw + checkOneValue + extractFirstToken +
    readFrameStorage + readFrameKeyList + dumpFrameKeys +
    collectFrameStorageValues + discoverFromAllFrames
  - StorageScanAll.ts (67 LoC) — Tier 3c: readAllJsonStorageValues +
    scanFrameForTokens + firstScanToken + discoverFromAllStorageKeys
  - PollTier.ts (72 LoC) — Tier 4: handleToToken + buildFrameWaiter +
    firstWaiterToken + pollForAuthModule
  - Orchestrator.ts (68 LoC) — storageTiers + fallbackTiers +
    discoverAuthThreeTier (default export)
  - index.ts (11 LoC) — EXPLICIT named re-exports
- Collapse old AuthDiscovery.ts to a 20-LoC deprecated shim re-exporting
  through ./AuthDiscovery/index.js (AUTH_HEADER_NAMES,
  discoverAuthThreeTier, discoverFromHeaders). All 4 historical callers
  compile unchanged.
- Every function in the new sub-modules is <= 10 effective LoC (drives
  toward §11 tightening in commit 7).
- Sub-modules import IDiscoveredEndpoint from NetworkDiscoveryTypes.js
  (not NetworkDiscovery.js) to avoid introducing new madge cycles.
- Tokens.ts exports prefixToken as PrefixedAuthToken brand to satisfy
  Rule #15 (no exported primitive returns).

Verified:
- npx tsc --noEmit ✅
- npx tsup ✅ (lib/index.cjs build succeeds)
- npx eslint Mediator/Network/AuthDiscovery* ✅ (clean)
- npx tsx detect-dead-code.ts ✅ (353 Pipeline files, all importers ≥1)
- npx tsx lint-and-validate.ts ✅ (Rule #15 clean after brand wrap)
- npx jest --testPathPatterns=AuthDiscovery ✅ (141/141 tests pass
  across 10 suites)
- npx jest --testPathPatterns=Mediator.Network ✅ (384/384 tests pass
  across 33 suites)
- npx madge --circular → 21 pre-existing cycles (no new cycles)
- docs-coverage gate ✅ (Phase 8.5a sub-module helpers added to
  .github/docs-coverage-allowlist.txt with per-symbol `# Reason:`
  comments — covers BOTH commit 1 (Fetch/*) and commit 2
  (AuthDiscovery/*) new public exports that are internal helpers
  composed by parent modules, not user-facing surface)
